+++ components-api/src/main/scala/pl/touk/nussknacker/engine/api/json/decoders/TypingResultDecoder.scala
@@ -4,6 +4,7 @@ import cats.data.NonEmptyList
 import io.circe._
 import pl.touk.nussknacker.engine.api.json.TypingType
 import pl.touk.nussknacker.engine.api.json.TypingType.{typeField, TypingType}
+import pl.touk.nussknacker.engine.api.typed.StandardTypesClasses
 import pl.touk.nussknacker.engine.api.typed.typing._
 import pl.touk.nussknacker.engine.api.typed.typing.DisplayStrategy.{DefaultDisplayStrategy, JsonDisplayStrategy}
 
@@ -109,9 +110,9 @@ class TypingResultDecoder(loadClass: String => Class[_]) {
   }
 
   private def tryToLoadClass(name: String, obj: HCursor): Decoder.Result[Class[_]] = {
-    if (name == Typed.KlassForArrays.getName) {
+    if (name == StandardTypesClasses.ArrayClass.getName) {
       // loading of array class causes Failed to load class [Ljava.lang.Object; with [Ljava.lang.Object;
-      Right(Typed.KlassForArrays)
+      Right(StandardTypesClasses.ArrayClass)
     } else {
       Try(loadClass(name)) match {
         case Success(value) => Right(value)